+++ front/components/assistant_builder/server_side_props_helpers.ts
@@ -42,7 +42,7 @@ import { getContentNodesForDataSourceView } from "@app/lib/api/data_source_view"
 import type { MCPServerViewType } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
 import { AppResource } from "@app/lib/resources/app_resource";
-import type { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
+import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import logger from "@app/logger/logger";
@@ -72,6 +72,34 @@ export const getAccessibleSourcesAndApps = async (auth: Authenticator) => {
     };
   });
 };
+
+// We are moving resource fetch to the client side. Until we finish,
+// we will keep this duplicated version for fetching actions.
+export const getAccessibleSourcesAndAppsForActions = async (
+  auth: Authenticator
+) => {
+  return tracer.trace("getAccessibleSourcesAndAppsForActions", async () => {
+    const accessibleSpaces = (
+      await SpaceResource.listWorkspaceSpaces(auth)
+    ).filter((space) => !space.isSystem() && space.canRead(auth));
+
+    const [dsViews, allDustApps, allMCPServerViews] = await Promise.all([
+      DataSourceViewResource.listBySpaces(auth, accessibleSpaces, {
+        includeEditedBy: true,
+      }),
+      AppResource.listByWorkspace(auth),
+      MCPServerViewResource.listBySpaces(auth, accessibleSpaces),
+    ]);
+
+    return {
+      spaces: accessibleSpaces,
+      dataSourceViews: dsViews,
+      dustApps: allDustApps,
+      mcpServerViews: allMCPServerViews,
+    };
+  });
+};
+
 
 export async function buildInitialActions({
   dataSourceViews,
