+++ front/pages/api/w/[wId]/builder/assistants/[aId]/actions.ts
@@ -1,8 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { buildInitialActions } from "@app/components/assistant_builder/server_side_props_helpers";
+import { buildInitialActions, getAccessibleSourcesAndAppsForActions } from "@app/components/assistant_builder/server_side_props_helpers";
 import type { AssistantBuilderMCPConfiguration } from "@app/components/assistant_builder/types";
-import { getAccessibleSourcesAndAppsForActions } from "@app/lib/api/assistant/action";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
